Revert "removed useless cast"
This reverts commit 60e8d6ce8b6c5300e96273d4ff8645cce52e3290.

+++ atlas/core/gamelist/parser.cpp
@@ -71,7 +71,7 @@ namespace gl
 			}
 			else
 			{
-				infos.f95_thread_id = ( match.captured( "f95_id" ).toULongLong() );
+				infos.f95_thread_id = static_cast< uint64_t >( match.captured( "f95_id" ).toULongLong() );
 			}
 		}
 		else
